% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_dups}
\alias{find_dups}
\title{check no duplicate observations}
\usage{
find_dups(df, vars)
}
\arguments{
\item{df}{a dataframe}

\item{vars}{character vector of variable names to group by}
}
\description{
\code{find_dups} checks that there are no duplicate observations for
each country-indicator-year combination
}
