% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_missing_vars}
\alias{find_missing_vars}
\title{check variables are present}
\usage{
find_missing_vars(df, vars)
}
\arguments{
\item{df}{a data frame}

\item{vars}{character vector of variable names}
}
\description{
\code{find_missing_vars} checks that the specified variables are present in a dataframe
}
