% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lengthen}
\alias{lengthen}
\title{convert UIS submission to long format}
\usage{
lengthen(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
\code{lengthen} converts a UIS data submission to wide format
}
