% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{parse_rules}
\alias{parse_rules}
\title{parse validation rules}
\usage{
parse_rules(path)
}
\arguments{
\item{path}{path to file}
}
\description{
\code{parse_rules} reads a yaml file of rules, checks that it is formatted correctly
and returns an object of class validate::validator if checks pass.
}
