% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_uis_data}
\alias{read_uis_data}
\title{read UIS data}
\usage{
read_uis_data(path)
}
\arguments{
\item{path}{path to file}
}
\description{
\code{read_uis_data} reads a supplied csv data submission  to a dataframe.
The function checks that the csv is in the required and that the minimum
necessary columns needed to validate the data are present:
namely CO_CODE, IND_ID, IND_YEAR, FIG, QAL and MAGN,
}
