% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_summary}
\alias{validate_summary}
\title{Summarise validation test}
\usage{
validate_summary(data, rules)
}
\arguments{
\item{data}{data frame to validate}

\item{rules}{an object of class validator (returned from \code{parse_rules})}
}
\description{
\code{validate_summary} is a wrapper around \code{validate::confront}.
Provided a data frame and rule set, it returns a validation
summary of passed and failed rules.
}
