% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{violating_obs}
\alias{violating_obs}
\title{show violating observations}
\usage{
violating_obs(data, rules, names = NULL)
}
\arguments{
\item{data}{data frame to validate}

\item{rules}{an object of class validator (returned from \code{parse_rules})}

\item{names}{names of the rules to evaluate}
}
\description{
\code{violating_obs} returns a dataframe fo violating observations
}
